requirements.txt
fastapi
uvicorn[standard]
websockets
gunicorn
python-telegram-bot==20.8
python-dotenv
aiofiles
